Update the spec test suite submodule
This commit updates the `tests/testsuite` submodule which is a copy of
the spec tests from upstream and the various proposals repositories. The
major change included here is the update to the lexing of the text
format to allow "reserved" tokens to have string tokens embedded within
them (in a sense). This doesn't affect any parsing in the crate, only
the tokenization of the input.

Additionally there are a few minor updates to error messages plus an
new exceptional case where the spec interpreter is printing "unknown
global" for defined globals which I'm not really sure what's going on.