Add the ability to share allocations during validation (#760)
This commit adds API support to the function validation portion of
`wasmparser` to share allocations amongst validation of different
functions. This is done by adding a `FuncToValidate` step which is prior
to creating a `FuncValidator` and creation of a `FuncValidator` happens
when it's fused with a new type, `FuncValidatorAllocations`. The
allocations can be retrieved by consuming the validator as well.

Locally in the benchmarks this yields a 5-15% performance increase
depending on the benchmark.

Closes #714