wasm-compose: implement a new `CompositionGraph` API.
This commit implements a new `CompositionGraph` API that allows for building a
composition with more flexibility than what the `ComponentComposer` API
supports.

The `CompositionGraph` is geared towards live-editing of a composition graph.

This commit also reimplements `ComponentComposer` using the composition graph.

The previous instance graph and the encoding implementation for it have been
removed.