Update `RequiredOptions` processing in `wit-component` (#335)
This commit updates how canoical ABI options are applied to
imports/exports, namely being more conservative about what is required.
For example the `Realloc` option is no longer necessary for any
arguments to imported functions, even if they contain strings. This
reflects how `Realloc` is only needed for incoming lists/strings to wasm
as opposed to outgoing ones.

A small test was added for this along with refactorings to existing
tests which show how the `realloc` option was dropped in many cases.
Internally this change was implemented by changing the `RequiredOptions`
enumeration to `bitflags!` representation with a split set of flags for
"what does this type contain transitively" to handle lifting/lowering
differently.