Minor changes to naming in wasmparser/wasm-encoder (#779)
* Minor changes to naming in wasmparser/wasm-encoder

I recently implemented a library which will translate from `wasmparser`
instructions to `wasm-encoder` instructions using the
`for_each_instruction!` macro to make it easier to convert between the
two. This PR represents some updates to make the process easier such as:

* Renames instruction variants to match in both `wasmparser` and
  `wasm-encoder`.
* Changes all single-payload instructions in `wasm-encoder` to uniformly
  be represented as a tuple enum variant instead of some as a tuple
  variant and some as a struct variant.
* Changes struct-variant names in `wasm-encoder` to match the names in
  `wasmparser`.
* Some struct-variant names in `wasmparser` are now more descriptive in
  isolation, for example `dst_mem` instead of `dst`.

Overall this is mostly stylistic changes but it made writing such a
macro much easier. I don't currently plan at this time to land the macro
in this repository but as part of the `wit-component` tool it may be
destined for this repository eventually which would provide a better
check that this is compatible into the future.

* Update tests for name changes, and make more changes

* Add `ReturnCall` and `ReturnCallIndirect` to `wasm-encoder`
* Rename `segment` fields to either `data_index` or `elem_index` to
  statically be able to tell which is which.
* Replace `wasmparser` to `wasm-encoder` instruction translation in
  `wasm-mutate` with a macro-based solution.
* Remove the `flags` field on `AtomicFence` in `wasmparser`, instead
  validating as part of the binary reading step that the flags are 0.