Leverage Cargo workspace deduplication
This commit is similar to bytecodealliance/wasmtime#4905 but for this
repository's family of crates. The goal here is to reduce the
version-complexity of the manifests in this repository by moving
dependencies specified in multiple places into the root `Cargo.toml` to
only specify versions once. This helps ensure that this family of crates
only use one version for all deps and additionally makes bumping
versions easier.

This also goes through and updates `edition` markers to all reference
the workspace so we can easily update the workspace of all crates in the
future as well.