Automatically validate required wasm proposals (#788)
This commit updates the operator validator within `wasmparser` to
automatically validate that the appropriate wasm proposal is enabled for
each instruction instead of manually listing the instructions. This is
done by using the `@proposal` tag in the `for_each_operator!` macro. A
wrapper visitor is used that's macro generated which validates the
required proposal before delegating to the actual validator which does
per-instruction validation.

This should also fix a few discrepancies where some instructions were
listed as `@reference_types` but validated as bulk-memory. Additionally
this aligned the `@`-names with the names in `WasmFeatures`.