Add support to wit-component to polyfill WASI  (#338)
* Add support to `wit-component` to polyfill WASI

This commit is an addition to the `wit-component` tool to be able to
polyfill WASI imports today using `wasi_snapshot_preview1` with a
component-model-using interface in the future. This is a large extension
to the functionality of `wit-component` internally since the generated
component is much "fancier".

The support in this commit is modeled as the addition of "adapter
modules" into the `wit-component` tool. An adapter module is understood
to translate from some core-wasm ABI into a component-model using ABI.
The intention is that for any previous API prior to the component model
an adapter module could be written which would translate from the prior
API to the new API. For example in WASI today there is:

    (@interface func (export "random_get")
      (param $buf (@witx pointer u8))
      (param $buf_len $size)
      (result $error (expected (error $errno)))
    )

whereas a component-model-using API would look more like:

    random-get: func(size: u32) -> list<u8>

This component-model version can be adapted with a module such as:

    (module $wasi_snapshot_preview1
      (import "new-wasi" "random_get" (func $new_random_get (param i32 i32)))
      (import "env" "memory" (memory 0))

      (global $last_ptr (mut i32) i32.const 0)

      (func (export "random_get") (param i32 i32) (result i32)
        ;; store buffer pointer in a saved global for `cabi_realloc`
        ;; later
        (global.set $last_ptr (local.get 0))

        ;; 1st argument: the `size: u32`
        local.get 1

        ;; 2nd argument: return pointer for `list<u8>`
        i32.const 8

        call $new_random_get

        ;; return a "success" return code
        i32.const 0
      )

      ;; When the canonical ABI allocates space for the list return value
      ;; return the original buffer pointer to place it directly in the
      ;; target buffer
      (func (export "cabi_realloc") (param i32 i32 i32 i32) (result i32)
        global.get $last_ptr)
    )

Using this adapter module the internal structure of the generated
component can be done such that everything is wired up in all the right
places meaning that when the original module calls
`wasi_snapshot_preview1::random_get` it actually calls this shim module
which then calls the actual `new-wasi::random_get` import. There's a few
details I'm glossing over here like the stack used by the shim module
but this suffices to describe the general shape.

My plan in the future is to use this support to generate a component
from all test cases that this repository supports. That means that,
specifically for `wit-bindgen` tests, a fresh new interface representing
"future WASI" will be created and the WASI functions used by tests will
be adapted via this adapter module. In this manner components will now
be generated for all tests and then the next step is #314, actually
ingesting these components into hosts.

* Update the gc translation macro

* Add gc comments

* Fix a missing `end` instruction on sp initializer

* Preserve the names of globals in adapter modules

Should help with debugging structure ideally

* Small refactor

* Hack in more stack pointer detection

This unfortunately suffers greatly from false negatives, but at this
time it's unclear if this can be done better.

* Improve readability of validation condition

* Improve validation documentation

* Comment a test case