Compile test cases with `wasm32-wasi` again
This commit is the next step in integrating `wasm32-wasi`,
`wit-bindgen`, tests, and components all together. Tests are now again
compiled with `wasm32-wasi` and use a repo-specific adapter module
(with support from #338) to support transforming the final module into
an actual component.

In supporting this feature the support from #331 is refactored into a
new `extract` Rust module so the functionality can be shared between
the ingestion of the main module as well as ingestion of adapter
modules. Adapter modules now are also supported on the CLI as a
standalone file without having to specify other options.

Note that the actual `wasi_snapshot_preview1.wasm` adapter is
non-functional in this commit and doesn't do anything fancy. The tests
in this repository don't really need all that much and I suspect all
we'll really need to implement is `fd_write` for fd 1 (as that's
stdout).