Fix wit-encoded exported interfaces with unused type aliases (#372)
Previously a panic would be hit during encoding since type type wasn't
encoded. This commit updates to encode all named types unconditionally
for exports to ensure this panic isn't hit.

Along the way this also removes an unnecessary `index_primitive` method
and map since during my prior refactorings in this code I found that it
was necessary to implement the functionality deeper and forgot to come
back and remove the original solution.