Implement the component-model lexing rules for identifiers. (#382)
* Implement the component-model lexing rules for identifiers.

The [component-model grammer] for kebab-case identifiers now looks
like this:

```
name           ::= <word>
                 | <name>-<word>
word           ::= [a-z][0-9a-z]*
                 | [A-Z][0-9A-Z]*
```

Implement the rules. This continues to use XID rules for the initial
lexing, as that corresponds to what users might accidentally use, so
that we can issue appropriate errors in those cases. The precise
grammer is validated in a separate step.

[component-model grammer]: https://github.com/WebAssembly/component-model/blob/main/design/mvp/Explainer.md#instance-definitions

* Add more lexing tests.

* Update the tests in tests/codegen/conventions.wit.

* Comment out identifiers that collide when mapped to snake_case, for now.

See https://github.com/WebAssembly/component-model/issues/118.