wasm-compose: simplify the example.
This commit simplifies the example by removing an unnecessary service layer
from the compose component.

Now the example either directly talks to a "service" component or composes a
new service component that injects a "middleware" component into the execution
graph.

It also updates the server in the example to Wasmtime 2.0.0.