wasmparser: Remove offset param in VisitOperator
Implementing `VisitOperator` is complicated by the additional
passing-through of the `offset` everywhere. A much simpler alternative
code wise is to stash the `offset` away within the implementation of
`VisitOperator` before invoking `SomeReader::visit_operator`.

This raises some complications with visitors that depend on the offsets
and rely on them for functional correctness, though. This is the case
for the `FuncValidator` which uses offsets to verify that the very last
`end` operator seen is at the same offset as the reader when the
validator is finalized.