Simplify API for parsing the name section
This commit modernizes the API to parse the `name` section in
`wasmparser`, notably leveraging existing traits and removing some
indirection through extra structures. The intent is to enable
iterator-based parsing similar to the rest of parsing items in this
crate, meaning the same idioms can be used for the `name` section as all
other sections.

At the same time this commit also adds a new subcommand I thought of
recently as `wasm-tools demangle` which runs a Rust or C++ demangler
over the `name` section. This is almost guaranteed to increase the size
of the `name` section but it otherwise might be useful when interacting
with tools that don't otherwise demangle names automatically, for
example web or JS runtimes.