wasmparser: rm `VisitOperator` for `FuncValidator`
This type depends on `offset`s being passed in correctly, that’s how it
validates whether the last `End` operator is actually the end of the
function.

Prior to removal of the `offset` parameters, they’d be correct by the
definition of `reader` passing the offset in. However, with move of this
responsibility to the user code, it now became straightforward to invoke
the visitor methods in a way that violates the assumptions in the
implementation of the `FuncValidator`.

Instead, in order to obtain a visitor, users are now required to call a
method which also sets up the offsets accordingly.