wasmparser: Remove offset param in VisitOperator (#804)
* wasmparser: Remove offset param in VisitOperator

Implementing `VisitOperator` is complicated by the additional
passing-through of the `offset` everywhere. A much simpler alternative
code wise is to stash the `offset` away within the implementation of
`VisitOperator` before invoking `SomeReader::visit_operator`.

This raises some complications with visitors that depend on the offsets
and rely on them for functional correctness, though. This is the case
for the `FuncValidator` which uses offsets to verify that the very last
`end` operator seen is at the same offset as the reader when the
validator is finalized.

* wasmparser: rm `VisitOperator` for `FuncValidator`

This type depends on `offset`s being passed in correctly, that’s how it
validates whether the last `End` operator is actually the end of the
function.

Prior to removal of the `offset` parameters, they’d be correct by the
definition of `reader` passing the offset in. However, with move of this
responsibility to the user code, it now became straightforward to invoke
the visitor methods in a way that violates the assumptions in the
implementation of the `FuncValidator`.

Instead, in order to obtain a visitor, users are now required to call a
method which also sets up the offsets accordingly.

* wasmparser: expand docs for `visit_operator`

* Do not run fuzzing for `cargo bench --all`