🏗️ Move wasm-smith's non-trapping mode logic to CodeBuilder (#769)
* Change CodeBuilder functions to emit multiple instructions

Co-authored-by: Nick Fitzgerald <fitzgen@gmail.com>

* Move no-trapping logic to a code_builder module

* Add configuration option to disallow traps

* Don't generate trapping instructions that haven't been handled yet
in non-trapping mode

* Remove unused import_count function

* Update crates/wasm-smith/src/core/code_builder.rs

* Add first pass at fuzzer to verify non-trapping mode does not trap

* rebase

* Address review comments

* Handle table and element segments in no-trapping mode

Co-authored-by: Nick Fitzgerald <fitzgen@gmail.com>

* Fix smoke test for non-trapping mode
It was occasionally running into an 'attempt to multiply with overflow'
for 32-bit memories, so we check the value before casting to avoid that.

* Clamp generated data offset to min memory size minus data length

* Handle running out of fuel during instantiation

* Don't generate simd instructions in non-trapping mode

* Use ne rather than eq to check for NaN values. NaN can't be checked
for using eq, but it is the only value that is not equal to itself,
so ne will work.

* Perform bounds checking for trunc instructions in non-trapping mode

* Adjust fuel to avoid "out of fuel" error on consume(0)

* Explicitly specify the largest/smallest i64 that can be represented as
an f64. This is necessary because `i64::MAX as f64` will result in an
f64 value that is not representable by an i64.

* As part of our non-trapping logic for load operations, we emit instructions
that must check if the memory offset plus the size of the type is within the bounds
of our memory. In order for us to emit this code, that value must fit in a u64.
In the case where the memory offset is u64::MAX, then our rust code panics in trying
to generate the wasm due to an attempt to add with overflow. To solve this, I've added
an upper bound on our memory offsets for load operations while running in non-trapping
mode.

* Don't emit unreachable instructions in non-trapping mode

* Explicitly specify max/min values for f32

* Handle min/max f32 values representable as i32

* Ensure stores/loads with negative addresses don't trap

* Don't allow stores with memarg.offset==i64::MAX while in non-trapping mode

* Refactor mem offset calculation to be in one place

* Use a ResourceLimiter to prevent libfuzzer from running out of memory.
This was happening with modules that had a large number of memory.grow
or table.grow instructions.

* Clean up code a little bit

* Add doc to `disallow_traps` explaining StackOverflow traps

Co-authored-by: Nick Fitzgerald <fitzgen@gmail.com>