Allow stack overflow in no-traps fuzzer
OSS-Fuzz found a case where a trap happened despite the fuel limits set
in place, which makes sense for this instance because it was a very
large function that blew the stack before the fuel limit was reached. In
general I don't think we can robustly protect against that so allow
stack overflow traps in the no-traps fuzzer.