wit-parser: create a top-level document parser.
This commit refactors `wit-parser` such that there is now a top-level
`Document` parser that can represent multiple worlds and interfaces in a single
wit document.

The previous internal `Document` type which represented the wit AST has been
renamed to simply `Ast`.

The `World::parse_file` method remains unchanged and will fail if there is not
exactly one world definition in the file.

A similar helper method `Interface::parse_file` was added back that fails to
parse if there is not exactly one interface definition in the file.

With these changes, the hack in the parsing tests to always have a world has
been removed.