Update Wasmtime and turn some features off
This is intended to fix build failures on oss-fuzz stemming from a
recent update to `zstd-sys` which includes the `-flto=thin` flag by
default that is apparently not compatible with the oss-fuzz builders.
This both updates Wasmtime and turns off the `cache` feature so the
`zstd-sys` crate is not pulled in.