Fix mutate fuzzer allowing too much fuel
Due to using the same store for each function invocation the allowance
of 1000 fuel kept getting accounted into the `fuel_consumed` calculation
since that is the amount of fuel consumed for the whole lifetime of the
store. This meant that each invocation of an exported function could get
twice as much fuel as the last. This commit fixes the issue by keeping
track of how much fuel the store has consumed and accounting for that
when giving the mutated function fuel.