wasmparser: split parsers for different elems (#849)
Given an `elem`, whether it is going to contain function indices or
constexprs is determined early on, and once determined, this won’t
change.

The way wasmparser’s API was structured suggested that such change
within a single element is in fact possible, which confused me every
time I encountered this. By introducing two distinct readers for each
kind of element items, this property is encoded in the API, much like it
is in wasm_encoder.

Although unverified and not at all the focus of this PR, this change may
also improve the performance of parsing items, as the branch checking
for whether the item is an expression of a function is now outside of
the loop by construction.