Add debug names to shim module in `wit-component` (#853)
I was debugging a runtime error recently and noticed that the backtrace
was missing function names for these shims generated by `wit-component`.
This commit adds a `name` section to assist in debugging. It's expected
that if this is undesired for file-size reasons then there will be other
build pipeline mechanisms to strip custom sections like the `name`
section.