Remove `deterministic` crate feature from `wasmparser (#854)
* Remove `deterministic` crate feature from `wasmparser

This is a very old feature of the `wasmparser` crate which I don't think
fits well within the current design of the crate. This is replaced with
a field in `WasmFeatures` that's always checked and always available to
work like other proposals. Additionally the `deterministic_only` field
in `WasmFeatures` has been inverted and renamed as `floats` to work like other
fields where enabling means allowing more behavior. This also
encompasses what the flag actually gates right now, which is just floats.

Closes #840

* Update CI actions