Try out a new style of section iterators
Almost all sections are a list of items and the code for managing this
is duplicated in quite a few places within the `wasmparser` crate. This
commit starts out a new abstraction, `SectionLimited`, with the goal
of eventually removing the `SectionReader` and `SectionReaderLimited`
traits entirely. The new lynchpin will be a `FromReader` trait
individually implemented for each type.