Consolidate list-reading functions into `read_iter`
This commit adds a new `BinaryReader::read_iter` method which is used to
consolidate all list-reading methods into one. This cleans up some
duplicate code in locations to read a size and then collect that many
items onto a returned list.

In the process this also adds `FromReader<'a> for ValType`.