Rewrite how types are managed in `wit-component`
This commit is a rewrite/refactor of how types are generated in the
final component as part of the `wasm-tools component new` command.
Previously types were all generated up-front and then referenced
afterwards, but this will not be compatible with `use` since not all
types can be generated first. For example a type which is used from
another interface needs to have the interface imported first to then
have the type available to alias out and use as part of another type.

The refactoring of this commit was largely to move the explicit type
generation to being interleaved with everything else being generated.
I've tried to split out some new files as well to cut down on the size
complexity of `encoding.rs`. Another minor change here is that this
restructuring makes it so only "live" types are printed now which are
those transitively reachable from functions that are exported.

Testing here is relatively light at the moment due to lack of `use`
parsing support in-tree.