Remove a large amount of boilerplate parsing from `wasmparser` (#857)
* Try out a new style of section iterators

Almost all sections are a list of items and the code for managing this
is duplicated in quite a few places within the `wasmparser` crate. This
commit starts out a new abstraction, `SectionLimited`, with the goal
of eventually removing the `SectionReader` and `SectionReaderLimited`
traits entirely. The new lynchpin will be a `FromReader` trait
individually implemented for each type.

* Remove `SectionReaders` in core wasm names

* Remove SectionReader usage from core exports

* Remove SectionReader from tag types

* Remove SectionReader from data section

* Remove SectionReader from memory section

* Remove SecitonReader from table section

* Remove SectionReader from core function section

* Remove SectionReader from core global section

* Remove SectionReader from code section parsing

* Remove SectionReader from canonical function section

* Remove SectionReader from core element section

* Remove SectionReader from core type section

* Remove SectionReader from core type section

* Remove SectionReader from component exports

* Remove SectionReader from producers section

* Remove SectionReader from component instances

* Remove SectionReader from component types and imports

* Remove reloc and linking section readers

I plan to re-add these back soon but otherwise they're quite outdated
and using old idioms so instead of updating their idioms just to get
rewritten later remove them for now to get re-added.

* Remove SectionReader from the component start section

This also updates the representation of the component start section to
have a single item instead of using a section reader of any form to
match the start section of core wasm.

* Remove SectionReader from core/component names

This required creating a new `Subsections` abstraction since these
custom sections aren't length-delineated like all the other prior sections.

* Remove `SectionReader` and related traits from wasmparser

Everything is now based on either `Subsections` or `SectionLimited`