Shrink `TypeId` back to 16-bytes
This commit preserves the 16-byte size of `TypeId` from `main` by
shrinking two fields back from `usize` to `u32`. The `type_size` field
was updated everywhere to use `u32` instead of `usize` since it never
exceeds 100k anyway and doesn't need `usize` on 64-bit platforms. The
`unique_id` integer was additionally shrunk to 32-bit as modules won't
be able to otherwise define 4B types.

Additionally an assertion is added in this commit to ensure that the
size of this structure stays relatively small.