Implement infrastructure fro a `use` keyword in `*.wit` (#858)
* Lift types in `wit-parser` to a `Document`

Additionally add `WorldId` and `InterfaceId` and start centralizing
around `Document` for the shared context type to be worked with. Many
helpers previously on `Interface` now live at the `Document` layer.

* Implement a strawman of merging documents/worlds

This is intended to be used by the `wit-component` CLI

* Update wit-component printing/decoding

This updates the decoding phase so factor in `use` and additionally
updates printing to print out `use` as well. Both are migrated to a
`Document` additionally.

* Get wit-component compiling again

* Get wit-component's interfaces test passing

* Get wit-component's tests working again

* Fix flattening of 0 flags

* Add simple wit fuzzer

* Fix representation of `use` types

Remove the superfluous `String` from the type list in an interface and
then change slightly how `use`-d types are represented. This ensure that
printing/encoding/decoding all agree on how these should appear and also
makes imported types usable as a `TypeId` for other items.

* Move raw component building to a separate file

Try to cut down on the size/complexity of `encoding.rs` by moving out
this piece to a dedicated builder file.

* Rewrite how types are managed in `wit-component`

This commit is a rewrite/refactor of how types are generated in the
final component as part of the `wasm-tools component new` command.
Previously types were all generated up-front and then referenced
afterwards, but this will not be compatible with `use` since not all
types can be generated first. For example a type which is used from
another interface needs to have the interface imported first to then
have the type available to alias out and use as part of another type.

The refactoring of this commit was largely to move the explicit type
generation to being interleaved with everything else being generated.
I've tried to split out some new files as well to cut down on the size
complexity of `encoding.rs`. Another minor change here is that this
restructuring makes it so only "live" types are printed now which are
those transitively reachable from functions that are exported.

Testing here is relatively light at the moment due to lack of `use`
parsing support in-tree.

* Fix issues with `use`

* In the fuzz generator perform a topological expansion of interfaces
  which will be done in resolution eventually but must be done by hand
  for now.

* Track live types from interfaces manually so when an interface has no
  exported functions but its types are required it's still emitted in
  the final component.

* Fix encoding `use`'d types in exported interfaces to alias the type
  properly.

* Fix detection of aliased/`use`'d types in decoding a `*.wit` from a
  `*.wasm`. This also required fixing a case in `wasmparser` to ensure
  types are given unique ids when aliased.

* Fix indexing with unused interfaces

* Merge wit-component tests into one

Have just one test function instead of multiple since nowadays they both
just panic for the same reason or both work.

* Assign truly unique ids in `TypeId`

Previously a collision was possible due to separate index spaces within
a component, so the type index could not be relied on for uniqueness.

* Add encoding support for aliases

Uses an outer alias of depth 0 to create a new type index for now,
this'll get replaced eventually with exported types creating a type index.

* Correct printing with type-aliases from components

Previously types would be visited in the wrong order and wasmparser/the
component don't have enough structure to report the hierarchy of a type
so visit types first to ensure it works for now. This ideally shouldn't
be possible in the future with exported types.

* Update wit-parser tests

Remove tests around the old `into_{world,document}` to get re-added
later if necessary, and otherwise update the json representation of a world.

* Fix named types with no interface in fuzzer

This isn't a valid AST, as all named types should have an interface
associated with them as the one that defined them.

* Remove cache in `index_of_type_export`

This fixes a fuzz-found issue where a type was aliased for an unrelated
interface but then the default export accidentally didn't assign its own
name to the type.

* Add some various comments to new code

* Fix fuzzer for now by commenting out failing cases

* Move wit fuzzing out of top-level fuzzing

This shouldn't run on oss-fuzz just yet.

* Shrink `TypeId` back to 16-bytes

This commit preserves the 16-byte size of `TypeId` from `main` by
shrinking two fields back from `usize` to `u32`. The `type_size` field
was updated everywhere to use `u32` instead of `usize` since it never
exceeds 100k anyway and doesn't need `usize` on 64-bit platforms. The
`unique_id` integer was additionally shrunk to 32-bit as modules won't
be able to otherwise define 4B types.

Additionally an assertion is added in this commit to ensure that the
size of this structure stays relatively small.

* Review comments, typos, etc

* Remove unnecessary set clone

* Remove trailing newlines in interfaces