wasmparser: simplify printing of shadowed labels (#866)
The webassembly spec test suite exercises label shadowing in the wat
format. `wasmprinter` would then utilize the annotations proposal to
maintain equivalent behaviour, but the reference interpreter does not
support it, which means that simply passing the spec tests through
`wasmparser` and `wasmprinter` would make tests unparseable by the
reference interpreter.

To fix this avoid the annotations extension for labels which support
shadowing.