Start full implementation of `use` for WIT
This commit is the first in what is likely to be a number of commits to
implement the full proposal for `use` in `wit-parser` and
`wit-component`. The immediate goal of this commit was to get a basic
structure in place for resolution and an AST for `wit-parser` to have.
To that end a new `UnresolvedPackage` container is added with a
`Resolve` container which can have unresolved packages added to it.

The main focus here is to get the `wit-parser` crate into a position
where most of its tricky bits are implemented and out of the way. This
includes resolution of multi-file packages, resolution of `use` within a
package, etc. This additionally includes new features such as
automatically topologically sorting everything inside containers and
full elaboration of world imports/exports with.

It's intended that the test suite of `wit-parser` is expanded to cover
all of this new functionality, errors and success cases. To that end the
tests here are:

* Existing tests all become "single file WIT packages". They are
  packages with one document and no sibling documents.
* A new test format, a directory, was added to the `tests/ui` folder
  which could house multi-file WIT packages which have optional package
  dependencies in a sub-`deps` folder.
* The JSON output of a successful parse was removed due to it getting
  unwieldy to continue updating and in general not being too
  interesting.

The `wit-component` crate is broken as a result of this commit, and the
next step is to update that to using `Resolve` primarily.