wasmprinter: add snapshot tests for the outputs
This PR saves the output of every wasmparser test into the
`tests/snapshot` directory. Once these are in place, every time
wasmprinter's output is changed, these changes will be reflected neatly
in these snapshots. This can aid with the detection of issues and also
as a mechanism to facilitate discussion of subjective calls on how
formatting should be done (i.e. previous formatting was better, or the
new formatting is better in this specific example.)

The downsite is that this is 2000 new files in the repo.