First pass at updating wit-component
Currently the crate compiles but most of it is commented out. This
commit specifically works on updating the printing portion as well as
the decoding portion from binary. Basic-to-semi-advanced binaries can be
decoded and printed now. No tests are currently written.

Printing is now centered around a document within a `Resolve`. The CLI
print currently requires that a document be specified if there's more
than one document. The plan is to also add something like `--out-dir` to
print out all documents at once.