Implement exports-introduce-indice semantics
This commit implements part of WebAssembly/component-model#129,
specifically the portion where an `export` item now introduces a fresh
item into the relevant index space. This is going to be used by #867 to
implement encoding/decoding of wit documents, specifically for `use`
statements.

This commit mostly implements the infrastructure necessary for this
change rather than any other major changes internally. Functionally
`export` statements now basically behave the same as `alias` statements
within the validator. Eventually more validation will be required to
ensure that functions use transitively-named types in all locations.
Additionally the name that a type index is exported at will be relevant
and will need exposing, but that'll come at a later date as well. It's
hoped that exposing this information, as required for WIT
encoding/decoding, will be a relatively small layer on the validator
from this commit.