Update the version number header of components
This commit performs the first update of the binary header version
number of the component binary encoding. The 4-byte header value, for
core WebAssembly, is 0x00000001. This 4-byte header value is proposed to
be reinterpreted as two 16-bit integers. The upper 16-bit integer of the
32-bit header version is reinterpreted as:

* 0x0000 - this is a core module
* 0x0001 - this is a component

The lower 16-bits are then the version within the format. For core wasm
the only supported version is 0x0001 and for components the previous
supported version was 0x000a (with the intention to move to 0x0001
on stabilization). This 0xa header value has never changed throughout
the lifetime of components but I would personally like to start using it
for breaking changes to the binary format during development. In the
spirit of that I've bumped it here to 0xb due to the change of how
`export` introduces an alias. This should break almost all preexisting
component binaries, indicating that a recompilation with newer tooling
is required. It's hoped that the error message produced here will be
more actionable than obscure indexing error messages.