Implement scaffolding of `wasm-tools component wit`
This commit implements the new `wasm-tools component wit` subcommand
which is a swiss-army-knife of working with:

* WIT textual documents as a single file-package
* WIT textual packages as a directory
* WIT packages encoded as a WebAssembly binary

Inter-conversion between all of these should now be available via all
the various flags and such on the CLI. This is all intended to help with
debugging, inspecting, and otherwise working with the low-level parts of
the component model. In other words this is trying to expose as much
functionality as possible implemented as libraries through a CLI for
when it's needed as a CLI is more quickly accessible than writing a
custom wrapper.