Implement roundtripping WIT through wasm
This commit implements all the necessary infrastructure to roundtrip a
WIT package through the WebAssembly component binary format. The
`wasm-tools component wit` subcommand is the tooling for this mode and
the `tests/interfaces` directory for `wit-component` has all been
updated. Tests are now either a single-file wit package or a directory
which is a wit-package. Tests have a `$name.wat` assertion for the
expected output and a `$name.wit.print` assertion for each document for
the machine-printed WIT as-decoded-from-wasm. The `tests/interfaces`
test also tests round-trip-ness to ensure that everything stays stable.

I've taken a number of liberties throughout the implementation of
encoding/decoding, primarily around the handling of URLs. Additionally
much of the encoding code related to creating an actual component from a
core wasm module is commented out here since I'll handle that in
subsequent commits. Tests were all updated here to the new syntax and
additionally all of the examples from the upstream component-model PR
have been added here as tests.