Rely less on the precise structure of URLs
This commit updates the WIT decoding process to not actually rely on the
structure of `pkg:/...` URLs at all, instead purely maintaining a hash
map of them as opaque identifiers. The purpose of this is to hopefully
future-proof against changes in the URL field which only requires that
they be unique.

This still relies on the URL structure, however, for foreign
dependencies. I'm not sure how to get around that at this time because
there's no other way to organize just imported interfaces back into
packages and documents.