Better handle `TypeId` and exported types
This commit updates the `ComponentEntityType::Type` variant to contain
both the `TypeId` for the type that was exported as well the identifier
for the exported type (or imported). This allows decoders like
`wit-component` to understand the chain from what got exported to the
id of the export itself to fully understand alias chains.