Update resolution of foreign deps
This commit updates how foreign deps are resolved during `Resolve::push`
to remove the restriction that the `unresolved.foreign_deps` field lists
documents in the same order as they are within the main arena. This will
be difficult to maintain during decoding and otherwise isn't all that
necessary.

Additionally this commit adds further asserts that an unresolved
package's items are indeed paritioned into remote and local parts since
much of the indexing code here relies on that.