Update the roundtrip-wit fuzzer
This commit is an overhaul and reimplementation for the `roundtrip-wit`
fuzzer in the `wit-component` package. This updates the AST generation
to instead generate a textual format with a string which is then parsed.
Various properties about round-tripping are then tested as well. about
this document.

Additionally this lifts the fuzzer into the top-level directory to run
on oss-fuzz since this has run long enough locally without bugs and I
think it would be good to suss out bugs sooner rather than later via
oss-fuzz.