Encode all types from imported interfaces in WIT
Previously only live types were encoded into the document encoding but
this meant that during the decoding process the order of types in
upstream documents might change due to the fact that a document would
use a subset and then the world would define the whole set in a
different order. This isn't a major issue but can cause problems for
round-tripping so this commit resolves the issue by encoding the entire
set of types of an upstream interface when imported into a document.