Register anonymous types more aggressively in decoding
This fixes a fuzz bug that was cropping up where a decoded document,
while structurally correct, accidentally contained duplicate anonymous
types for items which weren't duplicated in the original document. This
registration is unfortunately not testable with the current harnesses as
it only shows up when the decoded document is used, so for now I'm
leaving this to be fuzz-tested rather than CI-tested.