Fix a slew of fuzz bugs in wit-{parser,component} (#886)
* Refactor how paths are chosen in roundtrip-wit

None of these are showstopping bugs but this should help the rate of
rejection of fuzz test cases by ensuring we consider more inputs as
valid to generate a document in the end.

* Remove erroneous clearing of types

This was done before when there wasn't a separate interface generator
struct but now doing this means that no `use` paths will ever get
generated which is not intended. This fixes a big hole in the fuzzer
coverage where `use` statements basically weren't generated.

* Fix generation of `use` in the fuzzer to generate a `part`

Previously it would append directly to the output which while valid
excludes it from the shuffling logic to hopefully find more issues.

* Throw out fuzz cases with invalid worlds

Right now I think it would be relatively hard to reject worlds that end
up having implicit overlapping imports. Additionally I think it would be
pretty difficult to construct only valid worlds without overlapping
imports. To take an easy path out for now reject fuzz test cases that
produce invalid overlapping-import worlds.

* Fix round-tripping WIT with use-chains

Previously the decoding process didn't have enough information about
type aliases to determine what exactly an alias was pointing to, leading
everything to `use` from the definition rather than the path it was
originally written down as. This commit adds more infrastructure to
`wasmparser` to track aliases and what they refer to to reconstruct the
chain in WIT decoding and decode the same interface.

* Remove unnecessary declares in WIT printing

This is no longer needed with the new `Resolve` structure and ends up
accidentally actively causing issues if it stays, so remove it entirely.

* Fix types being used from exported interfaces

When looking up types look first in the export map and fall back to the
import map to ensure this matches resolution in wit-parser.

* Preserve type declaration order when printing

Don't reorder types from different `use` statements.

* Encode all types from imported interfaces in WIT

Previously only live types were encoded into the document encoding but
this meant that during the decoding process the order of types in
upstream documents might change due to the fact that a document would
use a subset and then the world would define the whole set in a
different order. This isn't a major issue but can cause problems for
round-tripping so this commit resolves the issue by encoding the entire
set of types of an upstream interface when imported into a document.

* Deduplicate import names in WIT encoding

Upstream interfaces may have the same name so when importing them into
one document the names may need to be disambiguated locally. Note that
with URL-based structuring this doesn't affect the structure of the
decoded document, it's just an implementation detail of the
wasm-encoding of WIT.

* Process transitive dependencies of imported interfaces

This fixes an issue with the other fix two commits prior where if a full
interface is encoded then its full set of types may pull in more
transitive dependencies which also need processing.

* Sort packages topologically during decoding

This commit fixes an issue where packages weren't always showing up in
their topological ordering after decoding from a wasm document.
In a wasm document all interfaces are in their topological order but
reconstructing the package ordering requires manual intervention due to
interfaces being discovered more-or-less on-the-fly.

* Register anonymous types more aggressively in decoding

This fixes a fuzz bug that was cropping up where a decoded document,
while structurally correct, accidentally contained duplicate anonymous
types for items which weren't duplicated in the original document. This
registration is unfortunately not testable with the current harnesses as
it only shows up when the decoded document is used, so for now I'm
leaving this to be fuzz-tested rather than CI-tested.

* Perform a more stable topological sort

This commit updates the topological sort performed during WIT resolution
to produce more deterministic orderings between items to assist with
round-tripping. Previously a document round-tripped through wasm would
have different dependency listings passed to the topological sort due to
elaboration discovering dependencies not explicitly written down in the
text format. This caused document orderings to differ slightly and
serializations also changed a bit.

The implementation here now uses a different algorithm for topological
sorting which automatically sorts all dependencies not only by their
topological ordering but also their name. This produces a more stable
sort regardless of the structure of the graph when new edges are drawn
implied by preexisting edges.

This is unfortunately pretty difficult to test so I'm mostly relying on
fuzzing to continue to find issues here.

* Fix test expectations from prior sorting commit

* Update crates/wasmparser/src/validator/types.rs

Co-authored-by: Peter Huene <peter@huene.dev>

* Update crates/wit-component/src/decoding.rs

Co-authored-by: Peter Huene <peter@huene.dev>

Co-authored-by: Peter Huene <peter@huene.dev>